02191953 Parsing done, operations next

+++ utils.c
@@ -6,7 +6,7 @@
 /*   By: katakaha <katakaha@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/17 18:04:44 by katakaha          #+#    #+#             */
-/*   Updated: 2026/02/17 18:05:21 by katakaha         ###   ########.fr       */
+/*   Updated: 2026/02/19 16:21:47 by katakaha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
  * @note Validate the string is a number before passing to atoi
  */
 
-int	my_atoi(const char *str)
+long	my_atoi(const char *str)
 {
 	int		i;
 	int		flag;
@@ -40,5 +40,6 @@ int	my_atoi(const char *str)
 		result = 10 * result + (str[i] - '0');
 		i++;
 	}
-	return ((int)(result * flag));
+	return (result * flag);
 }
+
